Add route and view for showing an individual book
Removed methods from Book and Chapter models to find a users book or chapters as their are built in methods for doing so.
Modified the slug methods in every model to effectively treat underscores like spaces.
Simplified the logic for handling which options to show to a user through the by_user, index, and new views along with refactoring the links shown to be more useful.
Added logic to check for current user to books index route, and new book route.
Added a check to find user by slug for the books by user view.