Add view and routes for creating new chapter
Made it so the login route will redirect you if you're already logged in.
Moved unfinished route and view for reordering a book to the books controller and view folder.
Reorganized the way the new user form gathers data to simplify the user creation process.
Signing upon now automatically logs a user in under their new account.
Added a section for showing rack flash errors to the book by user view.
Added links for managing chapters to the book edit view.
Updated the links in books show view to match the routes for chapter interactions.