Finish routes and view for editing chapters
Add outline for a route and empty view to modify existing contents of a chapter.
Changed the post and delete routes for chapters to use the ID's of book and chapters instead of title or ordinal, along with modifying forms to use the new paths.
Added note about chapter number length and contents in the views for creating and editing a chapter.
Added labels for book and chapters to show and write views.