Update routes so that their paths are RESTful
Lots of links are now broken but I have a plan to resolve this.
Removed logged_in? helper method, and replaced everywhere it was used with current_user helper.
Added before method to books and chapters controller to set the current user for all routes.
If searching by id the project now only uses the find method (rather than find_by).
The route for viewing a users books now refers to the user your looking up as author, and the view has been updated.
The route for editing a chapters contents now uses modify_content instead of just modify, and the view has been renamed the same thing.
Added an empty nav element to layout, going to test adding it.