Add nav bar and rework links through views and controllers
All links are now valid.
Removed links in the navbar from views.
Added links to the chapter views.
Improved chapter display in the books delete view.
The chapter delete view now properly displays the lack of content.
Added /delete to the end of the chapter delete route to clarify its purpose.
Added /edit to the end of the chapter edit route to resolve a routing conflict with the show route.
Moved the individual chapter show route to the bottom of the routes lists to reduce the chance of routing conflicts.