Deleting a book now actually deletes associated chapters
Book and Chapter show views now display their author (when not viewed by their author).
In order to make deleted chapters display in the book deletion view they are now saved to the @chapters variable prior to deletion.
Forgot to mention this in the last commit but I removed a number of .save method calls throughout the project as they were useless.
Made what was what clearer in the books show and delete views (exmaple: book titles are now labeled as such).
Added a link to the associated book for chapter views were it made sense.