Add find_books_chapter_by_ordinal helper so that chapter views will find the correct chapter
This new method replaces find_by using ordinal as this results in the first chapter created with a matching ordinal, and not the chapter associated with the book with the correct ordinal.
The unique_ordinal? and unique_name? methods have had their chapter variable names renamed to check_chapter, as before the argument chapter variable name didn't match the one in the method body.
Chapter and book deletion now uses the destroy method instead of delete.
The show views for books and chapters now show a link when you're not their author to more books by that author.
The chapter form fields for numbers have been converted from text back to number.
Corrected the form action link for modifying chapters content.
Added link to blog post.